refactor blogposts to its own file
