+++ src/handlers/talks.rs
@@ -51,7 +51,7 @@ pub async fn post_view(
                 .with_label_values(&[name.clone().as_str()])
                 .inc();
             let body = maud::PreEscaped(&post.body_html);
-            Ok((StatusCode::OK, tmpl::talk_view(&post, body, referer)))
+            Ok((StatusCode::OK, tmpl::blog::talk(&post, body, referer)))
         }
     }
 }
